Make azure adapter more aligned with vertex adapter
